Fix commit

+++ config/schema.sql
@@ -7,7 +7,7 @@ create table chatbots (
   created_at            datetime                                        comment 'Chatbot Creation Time',
   promptTemplate        text                                            comment 'Prompt Template',
   model                 varchar(30)         not null                    comment 'Model',
-  temperature           smallint            not null default 0          comment 'Temperature',
+  temperature           float               not null default 0          comment 'Temperature',
   visibility            varchar(30)         not null default 'public'   comment 'Private: No one can access your chatbot except you. Public: Anyone with the link can access it on chatbase.co and can be embedded on your website. Public: Anyone with the link can access it on chatbase.co and can be embedded on yoru website.',
   ip_limit              int(10)             not null default 20         comment 'Limit the number of messages sent from one device on the iframe and chat bubble',
   ip_limit_message      varchar(512)                                    comment 'Show this message to show when limit is hit',
